+++ ui/StructTraits.h
@@ -13,10 +13,10 @@ public:
 	typedef T value_type;
 	typedef T item_type;
 
-	static size_t size() { return 1; }
+    static int size() { return 1; }
 	static QStringList header() { return QStringList(); }
-	static const item_type get(const value_type& d, size_t /*i*/= 0) { return d; }
-	static void set(value_type& d, const item_type& v, size_t /*i*/= 0) { d = v; }
+    static const item_type get(const value_type& d, int /*i*/= 0) { return d; }
+    static void set(value_type& d, const item_type& v, int /*i*/= 0) { d = v; }
 };
 
 //***************************************************************//
@@ -28,14 +28,14 @@ public:
 	typedef QPointF value_type;
 	typedef qreal item_type;
 
-	static size_t size() { return 2; }
+    static int size() { return 2; }
 	static QStringList header()
 	{
 		QStringList header;
 		header << "X" << "Y";
 		return header;
 	}
-	static const item_type get(const value_type& d, size_t i = 0)
+    static const item_type get(const value_type& d, int i = 0)
 	{
 		switch(i)
 		{
@@ -45,7 +45,7 @@ public:
 
 		return 0.;
 	}
-	static void set(value_type& d, const item_type& v, size_t i = 0)
+    static void set(value_type& d, const item_type& v, int i = 0)
 	{
 		switch(i)
 		{
@@ -64,14 +64,14 @@ public:
 	typedef QRectF value_type;
 	typedef qreal item_type;
 
-	static size_t size() { return 4; }
+    static int size() { return 4; }
 	static QStringList header()
 	{
 		QStringList header;
 		header << "Left" << "Top" << "Right" << "Bottom";
 		return header;
 	}
-	static const item_type get(const value_type& d, size_t i = 0)
+    static const item_type get(const value_type& d, int i = 0)
 	{
 		switch(i)
 		{
@@ -83,7 +83,7 @@ public:
 
 		return 0.;
 	}
-	static void set(value_type& d, const item_type& v, size_t i = 0)
+    static void set(value_type& d, const item_type& v, int i = 0)
 	{
 		switch(i)
 		{
@@ -106,17 +106,17 @@ public:
 
 	enum { is_vector = 0 };
 	enum { is_single = 1 };
-	static size_t size(const vector_type&) { return 1; }
-	static const row_type* get(const vector_type& v, size_t i=0)
+    static int size(const vector_type&) { return 1; }
+    static const row_type* get(const vector_type& v, int i=0)
 	{
 		return !i ? &v : nullptr;
 	}
-	static void set(vector_type& v, const row_type& r, size_t i=0)
+    static void set(vector_type& v, const row_type& r, int i=0)
 	{
 		if(!i)
 			v = r;
 	}
-	static void resize(vector_type& /*v*/, size_t /*s*/) {}
+    static void resize(vector_type& /*v*/, int /*s*/) {}
 };
 
 //***************************************************************//
@@ -130,18 +130,18 @@ public:
 
 	enum { is_vector = 1 };
 	enum { is_single = 0 };
-	static size_t size(const vector_type& v) { return v.size(); }
+    static int size(const vector_type& v) { return v.size(); }
 	static QStringList header(const vector_type&) { return QStringList{}; }
-	static const row_type* get(const vector_type& v, size_t i=0)
+    static const row_type* get(const vector_type& v, int i=0)
 	{
 		return (i<size(v)) ? &(v[i]) : nullptr;
 	}
-	static void set(vector_type& v, const row_type& r, size_t i=0)
+    static void set(vector_type& v, const row_type& r, int i=0)
 	{
 		if(i < size(v))
 			v[i] = r;
 	}
-	static void resize(vector_type& v, size_t s)
+    static void resize(vector_type& v, int s)
 	{
 		v.resize(s);
 	}
